% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{epanechnikov}
\alias{epanechnikov}
\title{Epanechnikov kernel}
\usage{
epanechnikov(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A numeric vector of the Epanechnikov kernel evaluated at the
        values in \code{x}.
}
\description{
Vectorized evaluation of the Epanechnikov kernel.
}
