% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kre_adaptive.R
\name{nadwat}
\alias{nadwat}
\title{The Classical Nadaraya-Watson Regression Estimator}
\usage{
nadwat(x, dataX, dataY, K, h)
}
\arguments{
\item{x}{Numeric vector with the location(s) at which the Nadaraya-Watson
regression estimator is to be computed.}

\item{dataX}{Numeric vector \eqn{(X_1, \ldots, X_n)} of the x-values from
which (together with the pertaining y-values) the estimate is
to be computed.}

\item{dataY}{Numeric vector \eqn{(Y_1, \ldots, Y_n)} of the y-values from
which (together with the pertaining x-values) the estimate is
to be computed.}

\item{K}{A kernel function (with vectorized in- & output) to be used for the
estimator.}

\item{h}{Numeric scalar for bandwidth \eqn{h}.}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
In its arguments \code{x} and \code{dataX} vectorized function to compute
the classical Nadaraya-Watson estimator (as it is \eqn{m_n} in eq. (1.1)
in Eichner & Stute (2012)).
}
\details{
Implementation of the classical Nadaraya-Watson estimator as in eq. (1.1) in
Eichner & Stute (2012) at given location(s) in \code{x} for data \eqn{(X_1,
Y_1), \ldots, (X_n, Y_n)}, a kernel function \eqn{K} and a bandwidth \eqn{h}.
}
