% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jfunctions.R
\name{cuberoot}
\alias{cuberoot}
\title{Cube-root that retains its argument's sign}
\usage{
cuberoot(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Vector of same length and mode as input.
}
\description{
Computes \eqn{(x_1^(1/3), \ldots, x_n^(1/3))} with \eqn{x_i^(1/3)} being
negative if \eqn{x_i < 0}.
}
\examples{
kader:::cuberoot(x = c(-27, -1, -0, 0, 1, 27))
\donttest{
curve(kader:::cuberoot(x), from = -27, to = 27)
}

}
