% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{epanechnikov}
\alias{epanechnikov}
\title{Epanechnikov kernel}
\usage{
epanechnikov(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A numeric vector of the Epanechnikov kernel evaluated at the
        values in \code{x}.
}
\description{
Vectorized evaluation of the Epanechnikov kernel.
}
\examples{
kader:::epanechnikov(x = c(-sqrt(6:5), -2:2, sqrt(5:6)))
\donttest{
curve(kader:::epanechnikov(x), from = -sqrt(6), to = sqrt(6))
}

}
