% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convKfn_minimizeMSE.R
\name{kfn_vectorized}
\alias{kfn_vectorized}
\title{Convolution of Kernel Function K with fn}
\usage{
kfn_vectorized(u, K, xixj, h_sig)
}
\arguments{
\item{u}{Numeric vector.}

\item{K}{Kernel function with vectorized in- & output.}

\item{xixj}{Numeric matrix.}

\item{h_sig}{Numeric scalar.}
}
\value{
A vector of \eqn{(K * f_n)(u)} evaluated at the values in \code{u}.
}
\description{
Vectorized evaluation of the convolution of the kernel function K with fn.
}
\details{
Vectorized (in u) evaluation of - a more explicit representation of - the
integrand \eqn{K(u) * f_n(\ldots - h^2/\sigma * u)} which is used in the
computation of the bias estimator before eq. (2.3) in Srihera & Stute (2011).
Also used for the analogous computation of the respective bias estimator
in the paragraph after eq. (6) in Eichner & Stute (2013).
}
\note{
An alternative implementation could be
      \code{K(u) * sapply(h_sig * u, function(v) mean(K(xixj - v)))}
}
